# zhipin-kratos/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/position/{positionId}/chat-limited:
        post:
            tags:
                - CommunicationService
            operationId: CommunicationService_RecordChatLimited
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.RecordChatLimitedReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.CommunicationResp'
    /api/position/{positionId}/communications:
        get:
            tags:
                - CommunicationService
            operationId: CommunicationService_ListCommunications
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.ListCommunicationsResp'
    /api/position/{positionId}/interview:
        post:
            tags:
                - CommunicationService
            operationId: CommunicationService_RecordInterview
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.RecordInterviewReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.CommunicationResp'
    /api/position/{positionId}/msg-received:
        post:
            tags:
                - CommunicationService
            operationId: CommunicationService_RecordMsgReceived
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.RecordMsgReceivedReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.CommunicationResp'
    /api/position/{positionId}/msg-sent:
        post:
            tags:
                - CommunicationService
            operationId: CommunicationService_RecordMsgSent
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.RecordMsgSentReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.CommunicationResp'
    /api/position/{positionId}/rejection:
        post:
            tags:
                - CommunicationService
            operationId: CommunicationService_RecordRejection
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.RecordRejectionReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.CommunicationResp'
...
    /api/position/{positionId}/resume-sent:
        post:
            tags:
                - CommunicationService
            operationId: CommunicationService_RecordResumeSent
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.RecordResumeSentReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.CommunicationResp'
...
components:
    schemas:
...
        api.zhipin.CommunicationResp:
            type: object
            properties:
                id:
                    type: string
                positionId:
                    type: string
                eventType:
                    type: integer
                    format: int32
                eventTime:
                    type: string
                    format: date-time
                content:
                    type: string
                direction:
                    type: integer
                    format: int32
...
        api.zhipin.ListCommunicationsResp:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.zhipin.CommunicationResp'
...
        api.zhipin.RecordChatLimitedReq:
            type: object
            properties:
                positionId:
                    type: string
        api.zhipin.RecordInterviewReq:
            type: object
            properties:
                positionId:
                    type: string
                interviewTime:
                    type: string
                    format: date-time
                notes:
                    type: string
        api.zhipin.RecordMsgReceivedReq:
            type: object
            properties:
                positionId:
                    type: string
                content:
                    type: string
        api.zhipin.RecordMsgSentReq:
            type: object
            properties:
                positionId:
                    type: string
                content:
                    type: string
        api.zhipin.RecordRejectionReq:
            type: object
            properties:
                positionId:
                    type: string
                content:
                    type: string
        api.zhipin.RecordResumeSentReq:
            type: object
            properties:
                positionId:
                    type: string
                resumeVersion:
                    type: string
...